Move the effect to route layout
So that the "animation" is only between post types and taxonomies themselves
