+++ packages/js/src/settings/components/route-layout.js
@@ -1,7 +1,42 @@
+import { Transition } from "@headlessui/react";
+import { Fragment, useEffect } from "@wordpress/element";
+import { Title, usePrevious, useToggleState } from "@yoast/ui-library";
 import PropTypes from "prop-types";
 import { Helmet } from "react-helmet";
-import { Title } from "@yoast/ui-library";
+import { useLocation } from "react-router-dom";
 import { useDocumentTitle } from "../hooks";
+
+/**
+ * @returns {JSX.Element} The Route Transition Effect.
+ */
+const RouteTransitionEffect = () => {
+	const { pathname } = useLocation();
+	const previousPathname = usePrevious( pathname );
+	const [ hasRouteChanged, , , setHasRouteChangedTrue, setHasRouteChangedFalse ] = useToggleState( false );
+
+	useEffect( () => {
+		if ( pathname !== previousPathname ) {
+			setHasRouteChangedTrue();
+		}
+	}, [ pathname, previousPathname, setHasRouteChangedTrue ] );
+
+	return (
+		<Transition
+			as={ Fragment }
+			appear={ true }
+			show={ hasRouteChanged }
+			afterEnter={ setHasRouteChangedFalse }
+			enter="yst-transition-opacity yst-duration-150"
+			enterFrom="yst-opacity-0"
+			enterTo="yst-opacity-50"
+			leave="yst-transition-opacity yst-duration-150"
+			leaveFrom="yst-opacity-50"
+			leaveTo="yst-opacity-0"
+		>
+			<div className="yst-absolute yst-inset-0 yst-bg-white yst-select-none yst-rounded-lg" />
+		</Transition>
+	);
+};
 
 /**
  * @param {Object} props The properties.
@@ -28,6 +63,7 @@ const RouteLayout = ( {
 				</div>
 			</header>
 			{ children }
+			<RouteTransitionEffect />
 		</>
 	);
 };
